Extract Configuration Schema
Extract Schema out of Configuration.
This reduces the Configuration surface area, as well as prepares
for future features to be added to Configuration.
