+++ .rubocop.yml
@@ -9,10 +9,12 @@ AllCops:
 Metrics/ClassLength:
   Exclude:
     - "lib/google/gax/configuration.rb"
+    - "lib/google/gax/configuration/schema.rb"
 
 Style/Documentation:
   Exclude:
     - "lib/google/gax/api_call.rb"
+    - "lib/google/gax/configuration/schema.rb"
     - "lib/google/gax/errors.rb"
     - "lib/google/gax/grpc/status_details.rb"
 Style/GuardClause:
@@ -21,3 +23,4 @@ Style/GuardClause:
 Style/CaseEquality:
   Exclude:
     - "lib/google/gax/configuration.rb"
+    - "lib/google/gax/configuration/schema.rb"
